created note form to show up on dom. created event listener for saving notes
