+++ scripts/main.js
@@ -2,9 +2,10 @@ import { ConvictionSelect } from "./convictions/ConvictionSelect.js";
 import { CriminalList } from "./criminals/CriminalList.js";
 import { OfficerList } from "./officers/OfficerList.js";
 import { OfficerSelect } from "./officers/OfficerSelect.js";
+import { NoteForm } from "./notes/NoteForm.js";
 
-NoteForm()
 CriminalList()
 ConvictionSelect()
 OfficerList()
 OfficerSelect()
+NoteForm()
